Finish refactoring legacy js files
