+++ client/navigation/Attendance.tsx
@@ -1,48 +1,75 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View, Image } from 'react-native';
-import {NavigationContainer} from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import { NavigationContainer } from "@react-navigation/native";
+import { StatusBar } from "expo-status-bar";
+import { StyleSheet, Text, View, Image } from "react-native";
 
 export default function Attendance() {
   return (
     <View style={styles.container}>
-        
       {/* user profile icon image */}
-      <Image style={{ 
-          height: 50, 
-          width: 50, 
-          top: 0, 
+      <Image
+        style={{
+          height: 50,
+          width: 50,
+          top: 0,
           left: 140,
-        }} 
-        source={require('../assets/icon1.png')}
+        }}
+        source={require("../assets/icon1.png")}
       />
-    
+
       {/* Main Header */}
-      <Text style={styles.header}>Scan for {'\n'}attendance, {'\n'}userName</Text> 
+      <Text style={styles.header}>
+        Scan for {"\n"}attendance, {"\n"}userName
+      </Text>
 
       {/* Header for upcoming classes rectangle */}
       <View style={styles.bigSquare1}>
-        {<Text style={styles.classTimesText}>{'\t'}     Upcoming class times</Text>}
+        {<Text style={styles.classTimesText}>{"\t"} Upcoming class times</Text>}
       </View>
 
       {/* Rectangle displaying 1st upcoming class date and time */}
       <View style={styles.dateTimeRectangle1}>
-        {<Text style={styles.dateTimeText1}>Date: {'\t'}{'\t'}{'\t'}{'\t'} Time:</Text>}
+        {
+          <Text style={styles.dateTimeText1}>
+            Date: {"\t"}
+            {"\t"}
+            {"\t"}
+            {"\t"} Time:
+          </Text>
+        }
       </View>
 
       {/* Rectangle displaying 2nd upcoming class date and time */}
       <View style={styles.dateTimeRectangle2}>
-        {<Text style={styles.dateTimeText2}>Date: {'\t'}{'\t'}{'\t'}{'\t'} Time:</Text>}
+        {
+          <Text style={styles.dateTimeText2}>
+            Date: {"\t"}
+            {"\t"}
+            {"\t"}
+            {"\t"} Time:
+          </Text>
+        }
       </View>
 
       {/* Rectangle displaying 3rd upcoming class date and time */}
       <View style={styles.dateTimeRectangle3}>
-        {<Text style={styles.dateTimeText3}>Date: {'\t'}{'\t'}{'\t'}{'\t'} Time:</Text>}
+        {
+          <Text style={styles.dateTimeText3}>
+            Date: {"\t"}
+            {"\t"}
+            {"\t"}
+            {"\t"} Time:
+          </Text>
+        }
       </View>
 
       {/* QR code button (attendance) */}
       <View style={styles.clickHereRectangle}>
-        {<Text style={styles.clickHereText}>Click here to scan attendance QR code</Text>}
+        {
+          <Text style={styles.clickHereText}>
+            Click here to scan attendance QR code
+          </Text>
+        }
       </View>
 
       {/* Rectangle tracking total student hours amongst different sessions */}
@@ -50,30 +77,35 @@ export default function Attendance() {
         {
           // Date rectangle and header
           <View style={styles.dateRectangle}>
-          {<Text style={styles.dateRectangleText}>Date</Text>}
+            {<Text style={styles.dateRectangleText}>Date</Text>}
           </View>
         }
         {
           // Time rectangle and header
           <View style={styles.timeRectangle}>
-          {<Text style={styles.timeRectangleText}>Time</Text>}
+            {<Text style={styles.timeRectangleText}>Time</Text>}
           </View>
         }
         {
           // Total hours rectangle and header
           <View style={styles.totalHoursRectangle}>
-          {<Text style={styles.totalHoursRectangleText}> Total {'\n'}Hours</Text>}
+            {
+              <Text style={styles.totalHoursRectangleText}>
+                {" "}
+                Total {"\n"}Hours
+              </Text>
+            }
           </View>
         }
         {
           // List displaying total student hours amongst different sessions
           <View style={styles.listRectangle}>
-            {<View style={styles.listRectangleDivider1}/>}
-            {<View style={styles.listRectangleDivider2}/>}
+            {<View style={styles.listRectangleDivider1} />}
+            {<View style={styles.listRectangleDivider2} />}
           </View>
         }
-      </View> 
-    <StatusBar style="auto"/>
+      </View>
+      <StatusBar style="auto" />
     </View>
   );
 }
@@ -82,15 +114,15 @@ export default function Attendance() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: '#10579B',
-    alignItems: 'center',
-    justifyContent: 'center',
+    backgroundColor: "#10579B",
+    alignItems: "center",
+    justifyContent: "center",
   },
-  header: { 
-    color: 'white',
-    fontWeight: 'bold',
+  header: {
+    color: "white",
+    fontWeight: "bold",
     fontSize: 32,
-    position: 'absolute',
+    position: "absolute",
     top: 20,
     left: 20,
     margin: 16,
@@ -103,13 +135,13 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#5194D4',
+    backgroundColor: "#5194D4",
   },
   classTimesText: {
-    color: 'white',
-    fontWeight: 'bold',
+    color: "white",
+    fontWeight: "bold",
     fontSize: 18,
-    position: 'absolute',
+    position: "absolute",
     top: -10,
     left: -5,
     margin: 16,
@@ -122,13 +154,13 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#68B3FB',
+    backgroundColor: "#68B3FB",
   },
   dateTimeText1: {
-    color: 'white',
-    fontWeight: 'bold',
+    color: "white",
+    fontWeight: "bold",
     fontSize: 14,
-    position: 'absolute',
+    position: "absolute",
     top: -5,
     left: 0,
     margin: 16,
@@ -141,13 +173,13 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#68B3FB',
+    backgroundColor: "#68B3FB",
   },
   dateTimeText2: {
-    color: 'white',
-    fontWeight: 'bold',
+    color: "white",
+    fontWeight: "bold",
     fontSize: 14,
-    position: 'absolute',
+    position: "absolute",
     top: -5,
     left: 0,
     margin: 16,
@@ -160,13 +192,13 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#68B3FB',
+    backgroundColor: "#68B3FB",
   },
   dateTimeText3: {
-    color: 'white',
-    fontWeight: 'bold',
+    color: "white",
+    fontWeight: "bold",
     fontSize: 14,
-    position: 'absolute',
+    position: "absolute",
     top: -5,
     left: 0,
     margin: 16,
@@ -179,13 +211,13 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#68B3FB',
+    backgroundColor: "#68B3FB",
   },
   clickHereText: {
-    color: 'white',
-    fontWeight: 'bold',
+    color: "white",
+    fontWeight: "bold",
     fontSize: 15,
-    position: 'absolute',
+    position: "absolute",
     top: 0,
     left: 0,
     margin: 16,
@@ -198,7 +230,7 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#5194D4',
+    backgroundColor: "#5194D4",
   },
   dateRectangle: {
     width: 85,
@@ -209,13 +241,13 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#68B3FB',
+    backgroundColor: "#68B3FB",
   },
   dateRectangleText: {
-    color: 'white',
-    fontWeight: 'bold',
+    color: "white",
+    fontWeight: "bold",
     fontSize: 14,
-    position: 'absolute',
+    position: "absolute",
     top: -5,
     left: 10,
     margin: 16,
@@ -229,13 +261,13 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#68B3FB',
+    backgroundColor: "#68B3FB",
   },
   timeRectangleText: {
-    color: 'white',
-    fontWeight: 'bold',
+    color: "white",
+    fontWeight: "bold",
     fontSize: 14,
-    position: 'absolute',
+    position: "absolute",
     top: -5,
     left: 10,
     margin: 16,
@@ -249,13 +281,13 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 11,
     borderBottomLeftRadius: 11,
     borderBottomRightRadius: 11,
-    backgroundColor: '#68B3FB',
+    backgroundColor: "#68B3FB",
   },
   totalHoursRectangleText: {
-    color: 'white',
-    fontWeight: 'bold',
+    color: "white",
+    fontWeight: "bold",
     fontSize: 14,
-    position: 'absolute',
+    position: "absolute",
     top: -12.5,
     left: 7.5,
     margin: 16,
@@ -266,21 +298,21 @@ const styles = StyleSheet.create({
     top: -50,
     left: 22.5,
     borderWidth: 1,
-    borderColor: 'black',
-    backgroundColor: '#68B3FB',
+    borderColor: "black",
+    backgroundColor: "#68B3FB",
   },
   listRectangleDivider1: {
     width: 1,
     height: 108,
     top: 0,
     left: 90,
-    backgroundColor: 'white',
+    backgroundColor: "white",
   },
   listRectangleDivider2: {
     width: 1,
     height: 108,
     top: -108,
     left: 180,
-    backgroundColor: 'white',
+    backgroundColor: "white",
   },
 });
